refactor: Move Mui theme makeOverrides and makeShadows in helpers
